# Optimise the balance function