I stumbled upon it by accident and couldn't help myself. This is probably just a
drop in the bucket.

Microbenchmarks with an UTxO with 1,000,000 entries:

Before:

    time                 47.36 ms   (45.38 ms .. 48.76 ms)
                         0.997 R²   (0.994 R² .. 0.999 R²)
    mean                 47.86 ms   (47.23 ms .. 48.89 ms)
    std dev              1.582 ms   (1.131 ms .. 2.370 ms)

After:

    time                 18.72 ms   (18.53 ms .. 18.91 ms)
                         0.999 R²   (0.998 R² .. 1.000 R²)
    mean                 18.69 ms   (18.58 ms .. 18.86 ms)
    std dev              305.1 μs   (233.6 μs .. 431.3 μs)